+++ src/stability/fixedpoints.jl
@@ -1,6 +1,6 @@
 import IntervalRootFinding, LinearAlgebra
-using IntervalRootFinding: (..), (×), IntervalBox, interval
-export fixedpoints, .., ×, IntervalBox, interval
+using IntervalRootFinding: (..), (×), interval
+export fixedpoints, .., ×, interval
 
 """
     fixedpoints(ds::CoreDynamicalSystem, box, J = nothing; kwargs...) → fp, eigs, stable
@@ -12,11 +12,11 @@ Fixed points are returned as a [`StateSpaceSet`](@ref).
 For convenience, a vector of the Jacobian eigenvalues of each fixed point, and whether
 the fixed points are stable or not, are also returned.
 
-`box` is an appropriate `IntervalBox` from IntervalRootFinding.jl.
+`box` is an appropriate interval box from IntervalRootFinding.jl (a vector of intervals).
 E.g. for a 3D system it would be something like
 ```julia
-v, z = -5..5, -2..2   # 1D intervals, can use `interval(-5, 5)` instead
-box = v × v × z       # `\\times = ×`, or use `IntervalBox(v, v, z)` instead
+v, z = interval(-5, 5), interval(-2, 2)  # 1D intervals
+box = [v, v, z]
 ```
 
 `J` is the Jacobian of the dynamic rule of `ds`.
@@ -37,14 +37,14 @@ as a start of a continuation process. See also [`periodicorbits`](@ref).
   see the docs of IntervalRootFinding.jl for all possibilities.
 - `tol = 1e-15` is the root-finding tolerance.
 - `warn = true` throw a warning if no fixed points are found.
-- `order = nothing` search for fixed points of the n-th iterate of 
+- `order = nothing` search for fixed points of the n-th iterate of
   [`DeterministicIteratedMap`](@ref). Must be a positive integer or `nothing`.
   Select `nothing` or 1 to search for the fixed points of the original map.
 
 ## Performance notes
 
-Setting `order` to a value greater than 5 can be very slow. Consider using 
-more suitable algorithms for periodic orbit detection, such as 
+Setting `order` to a value greater than 5 can be very slow. Consider using
+more suitable algorithms for periodic orbit detection, such as
 [`periodicorbits`](@ref).
 
 """